fix: organize data structure by domain and fix directory nesting
- Each domain now has its own top-level folder within the data directory.
- Fixed an issue where folders were being nested incorrectly.
- Integrated project directory management into the Engine via DI.
- Verified all tests and coverage (85%) remain stable.

+++ uif_scraper/utils/circuit_breaker.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 
 class CircuitBreaker:
-    def __init__(self, threshold: int = 5, timeout: int = 300):
+    def __init__(self, threshold: int = 5, timeout: float = 300.0):
         self.threshold = threshold
         self.timeout = timeout
         self.failures: Dict[str, int] = {}
